+++ routes/web.php
@@ -119,16 +119,4 @@ Route::get('newprod', function(){
 
 //Forget Password
 
-//Cart Index
-Route::get('/cart', 'CartController@index')->name('cart.index');
-// Add item to cart
-Route::get('/addToCart/{id}', 'CartController@addToCart')->name('cart.add-item');
-// Minus quantity function
-Route::post('/updateItem', 'CartController@updateItem')->name('cart.update');
-// Remove Item
-Route::get('/removeItem/{id}', 'CartController@removeItem')->name('cart.remove');
-// Clear Cart
-Route::get('/clearCart', 'CartController@clearCart')->name('cart.clear');
-// Cart Checkout
-Route::post('/checkout', 'CheckoutController@checkout')->name('cart.checkout');
